Move Call.await into own file.
