+++ certificatetransparency/src/main/kotlin/com/babylon/certificatetransparency/internal/loglist/CallExt.kt
@@ -1,0 +1,60 @@
+/*
+ * Copyright 2019 Andrey Mischenko
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * File modified by Appmattus Limited to return a ByteArray or throw an exception
+ */
+
+package com.babylon.certificatetransparency.internal.loglist
+
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.suspendCancellableCoroutine
+import kotlinx.coroutines.withContext
+import okhttp3.Call
+import okhttp3.Callback
+import okhttp3.Response
+import java.io.IOException
+import kotlin.coroutines.resume
+import kotlin.coroutines.resumeWithException
+
+/**
+ * Suspend extension that allows suspend [Call] inside coroutine.
+ *
+ * @return Result of request or throw exception
+ */
+internal suspend fun Call.await(): ByteArray {
+    return withContext(Dispatchers.IO) {
+        suspendCancellableCoroutine { continuation ->
+            enqueue(object : Callback {
+                override fun onResponse(call: Call, response: Response) {
+                    continuation.resume(response.body()?.bytes() ?: throw IOException("No data available"))
+                }
+
+                override fun onFailure(call: Call, e: IOException) {
+                    // Don't bother with resuming the continuation if it is already cancelled.
+                    if (continuation.isCancelled) return
+                    continuation.resumeWithException(e)
+                }
+            })
+
+            continuation.invokeOnCancellation {
+                try {
+                    cancel()
+                } catch (ignored: Throwable) {
+                    // Ignore cancel exception
+                }
+            }
+        }
+    }
+}
